using our own power functions poly degree to int

+++ svm-train.c
@@ -159,7 +159,7 @@ void parse_command_line(int argc, char **argv, char *input_file_name, char *mode
 				param.kernel_type = atoi(argv[i]);
 				break;
 			case 'd':
-				param.degree = atof(argv[i]);
+				param.degree = atoi(argv[i]);
 				break;
 			case 'g':
 				param.gamma = atof(argv[i]);
